Added ability to simulate games wtih random moves
Adding this feature caught a bug where a peice moving backward onto a slide
would then bump itself off. This is because the way I implemented a slide is really
whack. The entire move/slide logic needs a revamp, because it is confusing and hacky right now.
I also think it could be made much more efficient.