Adding outputData flag. Updated Strategy
--outputData will now output CSV data detailing the game states and the winner.
I plan to feed this into a simple ML algorithm and see if I can train an AI.

I also added a simple update to Strategy that allows an AI to be configured to
make the move that best helps or hurts another player. Using this, if all 4 players
try to make another player win, then roughly 94% of the time that player will win.
Similarly, if all players try to make another player lose, then roughly 93% of the time
that player will lose. This is helpful knowledge for my games with my inlaws, where rivalries
and alliances are a critical part of the game.

I've also noticed the player 'B' seems to win more on average. I am wondering if this is because
going last is actually an advantage (due to the potential of more pawns on the board to Sorry) or
because of a bug in my hueristic computation.