Adding RandomForest to game simulator.
RF is *really* slow. Probably need to add parameters to the tree to make it smaller and faster.
Found a bug in the distance compute that treated home the same as safe:5. Fixed it.
Also found a bug in the UI that caused the safe spots to be treated as slides. Fixed it.